+++ user/src/main/kotlin/com/fx/user/application/service/FollowCommandService.kt
@@ -21,7 +21,6 @@ import org.springframework.stereotype.Service
 class FollowCommandService(
     private val followPersistencePort: FollowPersistencePort,
     private val profilePersistencePort: ProfilePersistencePort,
-    private val mediaWebPort: MediaWebPort
 ): FollowCommandUseCase {
 
     @Transactional
@@ -84,54 +83,9 @@ class FollowCommandService(
         return followPersistencePort.saveFollow(follow)
     }
 
-    override fun getUserFollowings(followQueryCommand: FollowQueryCommand): List<FollowUserInfo> {
-        checkFollowListAccess(followQueryCommand)
-
-        val followingUserList = followPersistencePort.getUserFollowings(
-            FollowQuery.searchCondition(followQueryCommand, FollowStatus.APPROVED)
-        )
-        
-        return mapMediaUrls(followingUserList)
-    }
-
-
-    override fun getUserFollowers(followQueryCommand: FollowQueryCommand): List<FollowUserInfo> {
-        checkFollowListAccess(followQueryCommand)
-
-        val followerList = followPersistencePort.getUserFollowers(
-            FollowQuery.searchCondition(followQueryCommand, FollowStatus.APPROVED)
-        )
-
-        return mapMediaUrls(followerList)
-    }
 
     override fun getFollowerByUserId(userId: Long): List<Long> =
         followPersistencePort.findByUserId(userId)
-
-
-    private fun checkFollowListAccess(followQueryCommand: FollowQueryCommand) {
-        val targetUserId = followQueryCommand.targetUserId
-        val requestUserId = followQueryCommand.requestUserId
-
-        // 자기 자신을 조회하는 경우 항상 허용
-        if (targetUserId == requestUserId) {
-            return
-        }
-
-        // 다른 사용자의 목록을 조회하는 경우
-        val targetUser = profilePersistencePort.findByUserId(targetUserId)
-            ?: throw ProfileException(ProfileErrorCode.PROFILE_NOT_FOUND)
-
-        // 대상 유저가 비공개 계정인 경우, 요청한 유저가 팔로우하고 있는지 확인
-        if (targetUser.isPrivate) {
-            val isFollowing = followPersistencePort.isFollowing(
-                requestUserId, targetUserId, FollowStatus.APPROVED
-            )
-            if (!isFollowing) {
-                throw FollowException(FollowErrorCode.UNAUTHORIZED_FOLLOW_LIST_ACCESS)
-            }
-        }
-    }
 
     private fun saveFollow(followerId: Long, followingId: Long, status: FollowStatus): Follow =
         followPersistencePort.saveFollow(
@@ -141,24 +95,4 @@ class FollowCommandService(
                 status
             )
         )
-
-    private fun mapMediaUrls(followUserInfoList: List<FollowUserInfo>): List<FollowUserInfo> {
-
-        val referenceIdList: List<Long> = followUserInfoList.mapNotNull { it.userId }
-
-        val mediaUrlMap: Map<Long, String> = if (referenceIdList.isNotEmpty()) {
-            mediaWebPort.getUrls(Context.PROFILE, referenceIdList)
-                .orEmpty()
-                .associate { it.referenceId to (it.mediaUrls?.firstOrNull().orEmpty()) }
-        } else {
-            emptyMap()
-        }
-
-        return followUserInfoList.map { user ->
-            user.copy(
-                profileImageUrl = user.mediaId?.let { mediaUrlMap[it] }
-            )
-        }
-    }
-
 }
